Rename view files
Files should not have the same name as their folder.
